backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment Variables
python-decouple==3.8

# CORS Headers
django-cors-headers==4.3.1

# Image Processing
Pillow==10.1.0

# File Upload
django-cleanup==8.0.0

# Filtering
django-filter==23.5

# API Documentation
drf-spectacular==0.27.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
